Refactor election lookup tools
