+++ static/js/pages/election-lookup.js
@@ -3,8 +3,8 @@
 /* global require, document */
 
 var $ = require('jquery');
-var lookup = require('../modules/election-lookup');
+var lookup = require('../modules/election-search');
 
 $(document).ready(function() {
-  new lookup.ElectionLookup('#election-lookup', true);
+  new lookup.ElectionSearch('#election-lookup', true);
 });
